+++ src/integration/c2b-payment-link/c2b-payment-link.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { SharedModule } from 'src/shared/shared.module';
-import { BinancePayService } from './binance-pay.service';
+import { BinancePayService } from './services/binance-pay.service';
 
 @Module({
   imports: [SharedModule],
   providers: [BinancePayService],
   exports: [BinancePayService],
 })
-export class BinancePayModule {}
+export class C2BPaymentLinkModule { }
